backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
PyJWT[crypto]==2.8.0
httpx==0.25.2
requests==2.31.0
supabase==2.3.3
gotrue==2.8.1
pydantic==1.10.15
python-dotenv==1.0.1
